// src/app/shared/models/EventPlayer.ts
export class EventPlayer {
    EventID: string;
    PlayerID: string;
    FactionID: string;
    ConflictChamberLink: string;
    ByVolunteerRounds: number[];
    DropRound: number;
    JoinRound: number;
    PairedDownRound: number;
    PairedUpRound: number;
    TP: number;
    CP: number;
    APD: number;
    Rank: number;

}